1. Fixed tools_def.template to meet ICC build for IA32 2. Modified some source files to meet ICC build for IA32 and IPF.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@3271 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Bus/Pci/UhciDxe/UhciSched.c
@@ -498,7 +498,7 @@ UhciCheckTdStatus (
       // next to the last known-good TD's data toggle if
       // any TD is executed OK
       //
-      QhResult->NextToggle = 1 - (UINT8)TdHw->DataToggle;
+      QhResult->NextToggle = (UINT8) (1 - (UINT8)TdHw->DataToggle);
 
       //
       // This TD is finished OK or met short packet read. Update the
